+++ knotx-knot/knotx-knot-action/src/main/java/com/cognifide/knotx/knot/action/FormConfigurationException.java
@@ -22,13 +22,7 @@ import com.cognifide.knotx.dataobjects.Fragment;
 
 public class FormConfigurationException extends RuntimeException {
 
-  private Fragment fragment;
-
   public FormConfigurationException(Fragment fragment) {
     super("Form not configured correctly " + fragment.knots());
-  }
-
-  public Fragment getFragment() {
-    return fragment;
   }
 }